requirements.txt
mysql-connector-python==9.1.0
flask==3.0.0
bootstrap-flask
python-dotenv==1.0.0 